+++ src/FluentDbTools/Abstractions/FluentDbTools.Common.Abstractions/IDbConfigCredentials.cs
@@ -1,0 +1,13 @@
+namespace FluentDbTools.Common.Abstractions
+{
+    public interface IDbConfigCredentials
+    {
+        string User { get; }
+
+        string Password { get; }
+
+        string AdminUser { get; }
+
+        string AdminPassword { get; }
+    }
+}