+++ src/FluentDbTools/Extensions/FluentDbTools.Extensions.Migration/ConfigureWithMigrationExtensions.cs
@@ -3,6 +3,8 @@ using System.Reflection;
 using FluentDbTools.Extensions.MSDependencyInjection;
 using FluentDbTools.Extensions.MSDependencyInjection.DefaultConfigs;
 using FluentDbTools.Common.Abstractions;
+using FluentDbTools.Extensions.Migration.DefaultConfigs;
+using FluentDbTools.Migration.Abstractions;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 
@@ -12,13 +14,21 @@ namespace FluentDbTools.Extensions.Migration
     {
         public static IServiceCollection ConfigureWithMigration(this IServiceCollection serviceCollection, IEnumerable<Assembly> assembliesWithMigrationModels)
         {
-            serviceCollection
+            return serviceCollection
                 .Register(FluentDbTools.Migration.Oracle.ServiceRegistration.Register)
                 .Register(FluentDbTools.Migration.Postgres.ServiceRegistration.Register)
+                .AddDefaultDbMigrationConfig()
                 .ConfigureWithMigrationAssemblies(FluentDbTools.Migration.ServiceRegistration.Register,
-                    assembliesWithMigrationModels)
-                .TryAddScoped<IDbConfig, MSDbConfig>();
-            return serviceCollection;
+                    assembliesWithMigrationModels);
         }
+
+        public static IServiceCollection AddDefaultDbMigrationConfig(this IServiceCollection serviceCollection)
+        {
+            return serviceCollection
+                .AddDefultDbConfig()
+                .AddScopedIfNotExists<IDbMigrationConfig, MsDbMigrationConfig>();
+        }
+
+
     }
 }