+++ src/FluentDbTools/Tests/Test.FluentDbTools.SqlBuilder.MinimumDependencies/StaticQueryBuilderInsertTest.cs
@@ -1,11 +1,11 @@
-﻿using FluentDbTools.Extensions.SqlBuilder;
-using FluentDbTools.Common.Abstractions;
-using TestUtilities.FluentDbTools;
+﻿using FluentDbTools.Common.Abstractions;
+using FluentDbTools.Contracts.DefaultConfigs;
+using FluentDbTools.Extensions.SqlBuilder;
 using FluentAssertions;
-using Test.FluentDbTools.SqlBuilder.TestEntities;
+using Test.FluentDbTools.SqlBuilder.MinimumDependencies.TestEntities;
 using Xunit;
 
-namespace Test.FluentDbTools.SqlBuilder
+namespace Test.FluentDbTools.SqlBuilder.MinimumDependencies
 {
     public class StaticQueryBuilderInsertTest
     {
@@ -18,7 +18,7 @@ namespace Test.FluentDbTools.SqlBuilder
         {
             var useSchema = !string.IsNullOrEmpty(schema);
 
-            var dbConfig = OverrideConfig.CreateTestDbConfig(databaseTypes, schema);
+            var dbConfig = DbConfigDatabaseTargets.Create(databaseTypes, schema);
             expectedSql = string.Format(expectedSql, dbConfig.Schema);
 
             var builder = dbConfig.CreateSqlBuilder();
@@ -44,7 +44,7 @@ namespace Test.FluentDbTools.SqlBuilder
             const string tableName = "EntityTable";
             var useSchema = !string.IsNullOrEmpty(schema);
 
-            var dbConfig = OverrideConfig.CreateTestDbConfig(databaseTypes, schema);
+            var dbConfig = DbConfigDatabaseTargets.Create(databaseTypes, schema);
             expectedSql = string.Format(expectedSql, dbConfig.Schema);
 
             var builder = dbConfig.CreateSqlBuilder();
@@ -70,7 +70,7 @@ namespace Test.FluentDbTools.SqlBuilder
         {
             var useSchema = !string.IsNullOrEmpty(schema);
 
-            var dbConfig = OverrideConfig.CreateTestDbConfig(databaseTypes, schema);
+            var dbConfig = DbConfigDatabaseTargets.Create(databaseTypes, schema);
             expectedSql = string.Format(expectedSql, dbConfig.Schema);
 
             var builder = dbConfig.CreateSqlBuilder();
@@ -85,7 +85,7 @@ namespace Test.FluentDbTools.SqlBuilder
 
             sql.Should().Be(expectedSql);
         }
-        
+
         [Theory]
         [InlineData(SupportedDatabaseTypes.Oracle, null, "INSERT INTO Entity(Id, Name, Description, EntityEnum, SomeProperty) VALUES(seq.nextval, 'Arild', :Description, :EntityEnum, :SomeProperty)")]
         [InlineData(SupportedDatabaseTypes.Postgres, null, "INSERT INTO Entity(Id, Name, Description, EntityEnum, SomeProperty) VALUES(nextval('seq'), 'Arild', @Description, @EntityEnum, @SomeProperty)")]
@@ -94,10 +94,8 @@ namespace Test.FluentDbTools.SqlBuilder
         public void InsertTest_WithEnumAndDirectProperty(SupportedDatabaseTypes databaseTypes, string schema, string expectedSql)
         {
             var useSchema = !string.IsNullOrEmpty(schema);
-
-            var dbConfig = OverrideConfig.CreateTestDbConfig(databaseTypes, schema);
+            var dbConfig = DbConfigDatabaseTargets.Create(databaseTypes, schema);
             expectedSql = string.Format(expectedSql, dbConfig.Schema);
-
             var builder = dbConfig.CreateSqlBuilder();
             var resolver = dbConfig.CreateParameterResolver();
             var insert = builder.Insert<Entity>();
@@ -109,7 +107,6 @@ namespace Test.FluentDbTools.SqlBuilder
                 .Fields(x => x.FP(f => f.EntityEnum))
                 .Fields(x => x.FP("SomeProperty"))
                 .Build();
-
             sql.Should().Be(expectedSql);
         }
     }